added functionality to add a node
This function can add a node, which can be either a file or folder.
It will first get current maximum ids, and set a this new node id to be maxId + 1.
Unit tested addNode() and findMaxId()

note that for better UI, files are added to the front of array, folders are to the back