added toggle open method in utils
previously we manage the isOpen status in each TreeNode component, but this won't allow user to provide custom open status. So we need to manage isOpen property at the top level tree state.

added toggle open method in utils and unit tested it so it will provide ability to manage isOpen at the top level.